Code cleanup, rearranged classes into packages
