+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="sarveshchavan777.quizgame">
+    package="blueteammathquiz.quizgame">
 
     <application
         android:allowBackup="true"
@@ -10,20 +10,20 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".LevelSelection.InstructionsActivity"></activity>
+        <activity android:name="blueteammathquiz.quizgame.segments.instructions.view.InstructionsActivity"></activity>
         <activity
-            android:name=".LevelSelection.PracticeLevelActivity"
+            android:name="blueteammathquiz.quizgame.segments.practice.view.PracticeLevelActivity"
             android:label="Practice" />
-        <activity android:name=".QuestionActivity">
+        <activity android:name="blueteammathquiz.quizgame.segments.game.view.QuestionActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
-        <activity android:name=".ResultActivity" />
+        <activity android:name="blueteammathquiz.quizgame.segments.game.view.ResultActivity" />
         <activity
-            android:name=".LevelSelection.SelectionLevelHomeScreen"
+            android:name="blueteammathquiz.quizgame.segments.home.view.SelectionLevelHomeScreen"
             android:label="Welcome!">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -32,7 +32,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".Splash"
+            android:name="blueteammathquiz.quizgame.segments.splash.view.SplashActivity"
             android:label="@string/app_name"
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
             <intent-filter>
@@ -42,7 +42,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".won"
+            android:name="blueteammathquiz.quizgame.segments.game.view.WonActivity"
             android:label="@string/app_name"
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen" />
     </application>
