+++ app/src/main/java/blueteammathquiz/quizgame/segments/game/model/Question.java
@@ -1,4 +1,4 @@
-package sarveshchavan777.quizgame;
+package blueteammathquiz.quizgame.segments.game.model;
 
 import android.app.Activity;
