+++ app/src/main/java/blueteammathquiz/quizgame/segments/game/view/ResultActivity.java
@@ -1,4 +1,4 @@
-package sarveshchavan777.quizgame;
+package blueteammathquiz.quizgame.segments.game.view;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -6,6 +6,8 @@ import android.os.Bundle;
 
 import android.view.View;
 import android.widget.TextView;
+
+import blueteammathquiz.quizgame.R;
 
 
 public class ResultActivity extends Activity {
