+++ app/src/main/java/blueteammathquiz/quizgame/segments/game/view/WonActivity.java
@@ -1,0 +1,33 @@
+package blueteammathquiz.quizgame.segments.game.view;
+
+import android.app.Activity;
+import android.os.Bundle;
+import android.widget.TextView;
+
+import blueteammathquiz.quizgame.R;
+
+/**
+ * Created by LENOVO on 1/3/2017.
+ * Modified by Nkechi Nnaji on October 13, 2019
+ */
+
+public class WonActivity extends Activity {
+    TextView tv;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_won);
+        tv = findViewById(R.id.congo);
+        Bundle b = getIntent().getExtras();
+        int y = b.getInt("score");
+
+        //Modified by Nkechi Nnaji
+        StringBuilder builder = new StringBuilder();
+        builder.append(getString(R.string.final_score)).append(" ").append(y);
+        tv.setText(builder);
+
+        TextView textView = findViewById(R.id.title);
+        textView.setText(R.string.congratulations); //Modified by Jayesh Shah
+    }
+}
