--- app/src/main/java/sarveshchavan777/quizgame/won.java
@@ -1,27 +1,0 @@
-package sarveshchavan777.quizgame;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.widget.TextView;
-
-/**
- * Created by LENOVO on 1/3/2017.
- */
-
-public class won extends Activity {
-    TextView tv;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.won);
-        tv = (TextView) findViewById(R.id.congo);
-        Bundle b = getIntent().getExtras();
-        int y = b.getInt("score");
-        tv.setText("FINAL SCORE:" + y);
-
-        TextView textView = (TextView) findViewById(R.id.title);
-        textView.setText(R.string.congratulations);
-    }
-}
