+++ app/src/main/res/layout/activity_splash.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-Splash layout modified by Nkechi Nnaji
+SplashActivity layout modified by Nkechi Nnaji
 -->
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
